set to only login user are allowed to like

+++ src/components/likes/Like.vue
@@ -1,6 +1,7 @@
 <template>
   <div>
     <img
+      :style="visitor ? 'cursor:not-allowed' : 'cursor:pointer'"
       class="icon-img"
       src="../../assets/icons/star.svg"
       data="star.svg"
@@ -13,13 +14,19 @@
 </template>
 
 <script>
-import { mapActions } from "vuex";
+import { mapActions, mapGetters } from "vuex";
 export default {
   name: "Like",
   props: {
     post_id: Number,
     likes: Array
   },
+  data() {
+    return {
+      visitor: 0
+    };
+  },
+  computed: mapGetters(["currentUser"]),
   methods: {
     ...mapActions([
       "addLike",
@@ -28,15 +35,20 @@ export default {
       "subtractPostLikes"
     ]),
     like() {
-      this.addLike({
-        id: this.post_id,
-        user_id: 1
-      })
-        .then(like => this.addPostLikes(like))
-        .catch(like =>
-          this.deleteLike(like.id).then(() => this.subtractPostLikes(like))
-        );
+      if (Object.keys(this.currentUser).length !== 0) {
+        this.addLike({
+          id: this.post_id,
+          user_id: this.currentUser.id
+        })
+          .then(like => this.addPostLikes(like))
+          .catch(like =>
+            this.deleteLike(like.id).then(() => this.subtractPostLikes(like))
+          );
+      }
     }
+  },
+  created() {
+    this.visitor = Object.keys(this.currentUser).length === 0 ? 1 : 0;
   }
 };
 </script>
